+++ src/cleaner.rs
@@ -24,12 +24,10 @@ use std::fs;
 use std::io::{Read, Write};
 use std::io;
 
-use clap::ArgMatches;
-
 use svgdom;
 use svgdom::{Document, ParseOptions, WriteOptions, WriteBuffer, ElementId};
 
-use cli::{Key, get_flag};
+use options::Options;
 use task::*;
 use error;
 
@@ -47,7 +45,7 @@ pub fn parse_data(data: &[u8], opt: &ParseOptions) -> Result<Document, svgdom::E
     Document::from_data_with_opt(data, opt)
 }
 
-pub fn clean_doc(doc: &Document, args: &ArgMatches, opt: &WriteOptions)
+pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
                  -> Result<(), error::Error> {
     try!(preclean_checks(doc));
 
@@ -66,137 +64,137 @@ pub fn clean_doc(doc: &Document, args: &ArgMatches, opt: &WriteOptions)
     // Do not remove any attributes before this methods
     // since they uses them.
 
-    if get_flag(args, Key::RemoveTitle) {
+    if options.remove_title {
         remove_element(doc, ElementId::Title);
     }
 
-    if get_flag(args, Key::RemoveDesc) {
+    if options.remove_desc {
         remove_element(doc, ElementId::Desc);
     }
 
-    if get_flag(args, Key::RemoveMetadata) {
+    if options.remove_metadata {
         remove_element(doc, ElementId::Metadata);
     }
 
-    if get_flag(args, Key::RemoveUnusedDefs) {
+    if options.remove_unused_defs {
         remove_unused_defs(doc);
     }
 
-    if get_flag(args, Key::RemoveInvalidStops) {
+    if options.remove_invalid_stops {
         remove_invalid_stops(doc);
     }
 
-    if get_flag(args, Key::ApplyTransformToGradients) {
+    if options.apply_transform_to_gradients {
         // Apply transform to gradients before processing to simplify duplicates
         // detecting and merging.
         apply_transforms::apply_transform_to_gradients(doc);
     }
 
-    if get_flag(args, Key::RemoveDuplLinearGradients) {
+    if options.remove_dupl_linear_gradients {
         remove_dupl_linear_gradients(doc);
     }
 
-    if get_flag(args, Key::RemoveDuplRadialGradients) {
+    if options.remove_dupl_radial_gradients {
         remove_dupl_radial_gradients(doc);
     }
 
-    if get_flag(args, Key::RemoveDuplFeGaussianBlur) {
+    if options.remove_dupl_fe_gaussian_blur {
         remove_dupl_fe_gaussian_blur(doc);
     }
 
-    if get_flag(args, Key::MergeGradients) {
+    if options.merge_gradients {
         merge_gradients(doc);
     }
 
-    if get_flag(args, Key::ApplyTransformToGradients) {
+    if options.apply_transform_to_gradients {
         // Do it again, because something may changed after gradient processing.
         apply_transforms::apply_transform_to_gradients(doc);
     }
 
-    if get_flag(args, Key::ApplyTransformToShapes) {
+    if options.apply_transform_to_shapes {
         // Apply before 'convert_shapes_to_paths'.
         apply_transforms::apply_transform_to_shapes(doc);
     }
 
-    if get_flag(args, Key::ConvertShapes) {
+    if options.convert_shapes {
         convert_shapes_to_paths(doc);
     }
 
     // NOTE: run before `remove_invisible_elements`, because this method can remove all
     //       segments from the path which makes it invisible.
-    if get_flag(args, Key::PathsToRelative) {
+    if options.paths_to_relative {
         // we only process path's segments if 'PathsToRelative' is enabled
-        paths::process_paths(doc, args);
+        paths::process_paths(doc, options);
     }
 
-    if get_flag(args, Key::RemoveInvisibleElements) {
+    if options.remove_invisible_elements {
         remove_invisible_elements(doc);
     }
 
-    if get_flag(args, Key::RegroupGradientStops) {
+    if options.regroup_gradient_stops {
         regroup_gradient_stops(doc);
     }
 
-    if get_flag(args, Key::UngroupGroups) {
+    if options.ungroup_groups {
         ungroup_groups(doc);
     }
 
-    if get_flag(args, Key::ResolveUse) {
+    if options.resolve_use {
         resolve_use(doc);
     }
 
     // now we can remove any unneeded attributes
 
-    if get_flag(args, Key::RemoveDefaultAttributes) {
+    if options.remove_default_attributes {
         remove_default_attributes(doc);
     }
 
-    if get_flag(args, Key::RemoveTextAttributes) {
+    if options.remove_text_attributes {
         remove_text_attributes(doc);
     }
 
-    if get_flag(args, Key::RemoveNeedlessAttributes) {
+    if options.remove_needless_attributes {
         remove_needless_attributes(doc);
     }
 
-    if get_flag(args, Key::RemoveGradientAttributes) {
+    if options.remove_gradient_attributes {
         remove_gradient_attributes(doc);
     }
 
-    if get_flag(args, Key::RemoveUnusedCoordinates) {
+    if options.remove_unused_coordinates {
         remove_unused_coordinates(doc);
     }
 
     // Run only after attributes processed, because
     // there is no point in grouping default/unneeded attributes.
-    if get_flag(args, Key::GroupByStyle) {
+    if options.group_by_style {
         group_by_style(doc);
     }
 
     // final fixes
     // list of things that can't break anything
 
-    if get_flag(args, Key::RemoveUnreferencedIds) {
+    if options.remove_unreferenced_ids {
         remove_unreferenced_ids(doc);
     }
 
-    if get_flag(args, Key::TrimIds) {
+    if options.trim_ids {
         trim_ids(doc);
     }
 
-    if get_flag(args, Key::RemoveVersion) {
+    if options.remove_version {
         remove_version(doc);
     }
 
-    if get_flag(args, Key::UngroupDefs) {
+    if options.ungroup_defs {
         ungroup_defs(doc);
     }
 
     remove_empty_defs(doc);
-    fix_xmlns_attribute(doc, get_flag(args, Key::RemoveXmlnsXlinkAttribute));
+    fix_xmlns_attribute(doc, options.remove_xmlns_xlink_attribute);
 
     // NOTE: must be run at last, since it breaks the linking.
-    if get_flag(args, Key::JoinStyleAttributes) {
+    if options.join_style_attributes {
         join_style_attributes(doc, opt);
     }