+++ src/lib.rs
@@ -20,10 +20,20 @@
 **
 ****************************************************************************/
 
+#[cfg(feature = "cli-parsing")]
 #[macro_use] extern crate clap;
+
 extern crate svgdom;
 
+pub use svgdom::{ParseOptions, WriteOptions};
+
+pub use options::Options as CleaningOptions;
+pub use error::Error;
+
+#[cfg(feature = "cli-parsing")]
 pub mod cli;
+
 pub mod cleaner;
-pub mod error;
+mod error;
+mod options;
 mod task;